msm: mdp: Fix mdp_hist_lut_write_off frame size
  * Fixes following error :
      drivers/video/msm/mdp.c: In function 'mdp_hist_lut_write_off':
      /drivers/video/msm/mdp.c:399:1: warning: the frame size of
        1032 bytes is larger than 1024 bytes [-Wframe-larger-than=]

+++ drivers/video/msm/mdp.c
@@ -387,7 +387,7 @@ static int mdp_hist_lut_write_off(struct mdp_hist_lut_data *data,
 	mdp_pipe_ctrl(MDP_CMD_BLOCK, MDP_BLOCK_POWER_ON, FALSE);
 	for (i = 0; i < MDP_HIST_LUT_SIZE; i++) {
 		last_lut[i] = element[i];
-		MDP_OUTP(MDP_BASE + base + offset + (0x400*(sel)) + (4*i),
+		MDP_OUTP((uint32_t)MDP_BASE + base + offset + (0x400*(sel)) + (4*i),
 				element[i]);
 	}
 	mdp_pipe_ctrl(MDP_CMD_BLOCK, MDP_BLOCK_POWER_OFF, FALSE);
